Class responsibilities segregated
